requirements.txt
transformers
torch
Pillow
requests
pandas





